Fixes https://github.com/Fenoxo/sourceTiTS/issues/14
I have no idea what `if(this.userInterface.leftSideBar.currentFrame >= 11)` evaluates for, but just removing the conditional check entirely seems to make everything work correctly?

The only thing I can think of is that now clicking the "data" button repeatedly toggles the save/load screen on and off, but that seems to be appropriate behaviour, so.... wut?