Rigged up some *reasonable* connections from game state -> gui for NPC secondary stats/player "advancement" stats
XP calcs need to be done for level up requirements.
NPCs need the Race() method to be built I guess? Are there going to be NPCs that transform, or do we just care about originalRace (in which case, we need a version of originalRace that's formatted for "Pretty Display" ie Cunt Snake vs cunt snake)
XP StatBar might need to be specialised for display purposes; the "placeholder" text was "<xp> / <xp_to_level>" but the way the statbars are built would require specialisation to display the text that way.