I think this fixes a few issues with CuntSnakes.
"Type" randomisation was being overriden in the method to proper a cuntsnake for combat. The descriptive text was being set based off of randomisation, and then the type (which would be called by tailVaginaDescript) was changed to something else.
I believe an inadvertent call to vaginaDescript which should have been tailVaginaDescript