Saendra polished and implemented. Fight works, might need some balancing and some tweaking of the combat abilities (I'm shit at figuring out where new lines should go in the combat stuff).
Basic placeholder "travel event" system - check travelEvents.as
Basically, there's a 25% every time your ship travels between locations to trigger a message. If it triggers ONE of the potential travel events will be selected and a message presented to the player.
The message itself is a wrapper to hide the specific event that you proc - the idea being, we don't end up with a "distress signal message" that hangs around for 3 months or whatever. Travel -> event proc -> "Do you want to read message? Y/N" -> {ACTUAL content message/ignore carry on with shit} DOIT/IGNORE -> {start content/content disabled or w/e}
If you don't want to deal with stuff, ignore the message and you're none the wiser - if you ignore the ACTUAL distress message, the content it's tied to is then disabled or possibly resolves some other way without the players involvement. Assuming the message is time sensitive (like a distress call).

Caught a couple of missing brackets in the genericApplyDamage function.
Tweaked the parsers quote replacement regexp (again). :|