Fixed potential vag shrinking bug?
This really looks like a bug to me, but I don't know enough about the mechanic to be sure. Also I don't own Flash Pro, so I can't compile or test it...

If it isn't a bug, I would recommend refactoring the code to be more expressive or throw a comment in there.