Resolves an issue introduced with Fall of the Phoenix between game save versions.
The players desired flight location before the message is introduced to them is stored in a flag. That flag is then later used to determine where the players ship should be sent once the quest is over. It also didn't exist in the initial release version, so anybody who saved mid-way through the quest doesn't have that flag set, and the game attempts to flyTo(undefined) rather than a valid location.