Updates to the sublime scripts.
Added a doubledash replacer - strips "--" and replaces it with a single dash of the correct type.
Added a remover for what has been dubbed "savvy spacing"- trailing spaces at the end of a line will be removed automatically.
Changed the regexp used to find and insert italics tags. The script will now automatically search for, and position <i> tags outside of fancyquotes according to the massive change the other day. Opening tag is before the quote, closing tag is after the quote.