Split the FLA into two versions to support ImagePack enabled and ImagePack disabled. Modified Flahnes Under Desk scene as the test bed for the image embedding system.
Doesn't quite work the same as the bust display. The images have to be wrapped in a child class that will handle resizing correctly when embedded into the text area. I'll be extending this for other things tomorrow.

Basic order of business to add a new image to the pack:
1. Place image in the appropriate assets folder. assets/images/imagepack/<npc>/<image>
2. Add the new embedded asset into classes/resources/imagepack.as:
[Embed(blah)]
public static const MyTestImageT:Class (T for Type effectively)
3. Copy one of the existing classes available under classes/resources/imagepack/ and rename it.
CtrlC+CtrlV -> Rename -> MyTestImageC.as, rename constructor etc etc to match. (Effectively C == ClassDef)
4. Change the asset targetted in the constructor of the newly copied class from the core ImagePack thing. Check out Flahne does it.
bmp = new ImagePack.MyTestImageT();
5. Go back to the root image pack AS file. Add the class definition pointing to the class you just made. Use the actual name you want here.
public static const MyTestImage:Class = MyTestImageC;

You should have 3 lines in ImagePack.as for the Embed, a MyClassT and a MyClass = MyClassC, and then a class under the actual ImagePack folder for MyClassC.