Wired some stuff together for damage output.
All damage outputting is now handled by a single function: outputDamage()
Pass it a damage result from an applyDamage or calculateDamage call, and it will appropriately format the result.
Apply damage already uses this, and will call it automatically, assuming applyDamage is not passed the "suppress" option
There's options in the options menu to configure it a bit: toggle "old mode" aka combined damage and color on/off

I've probably missed some attacks getting wired to this in various places. Basically, everything should move to using applyDamage/calculateDamage and doing things with the returned DamageResult for output.