Fixes the "0 eggs" thing.
Basically, the message was from a special variant of the sex scene for players to get multi-pregged (already have nyrea eggs, but still have an available pregnancy slot). It was tripping the checks to initiate that scene, but the pregnancy itself wasn't applying because multiPregnancy of those types was disabled. Ergo, it was correctly finding the hole, correctly attempting to impregnating, failing to do anything, and then getting the "empty slot" data of 0 quantity.

I think I disabled multipregs for something else and then later cleaned up the bug (because it was a lot of twiddling around to get it handled). It MIGHT still need some tweaking to unify the message generation across all the different pregnancy types to tone it down some so we're not flooding the player with messages.