Space Skype Scene Tweak
If the suit is nowhere to be found on the character or in the ship's inventory, how does this scene happen? (It is possible to sell her away to another vendor for 0 credits and she is not flagged as a quest item to protect from that.) Maybe add an item check for it!