Minor Corrections
Potential fix for: http://fenoxo.com/forum/index.php?/topic/461-065-briha-child-bug/

Looks like there was a confusion with child age and incubation time. If the incubation time it before 120 hours, the child is born as an egg, else it's been hatched. I think this is how it's supposed to work.