Minor Tweaks
Made the icons consistent. 
Beds, if room is not occupied by an NPC, get an objective marker (like some resources/camps do, I guess).
Staying in the prison will sometimes spawn a cave-dwelling enemy... but the space is not labelled as a hazardous area. I commented it out just in case.