Fix for displayInput() Bug
Potential fix for: http://fenoxo.com/forum/index.php?/topic/733-tits_0610-creation-and-data-management/

Basically, just added a dataOff() function and added it to the displayInput() so the player can't load a file while the text display is on--otherwise, it will cause a floating input bug throughout the game. Hopefully this fixes the issue.