Minor Tweaks
Nonesuch says: "That whole check is just to make sure the PC is wearing ANY lower undergarment - so that particular scene can make sense. It shouldn't matter whether it's swimwear or anything else."

So I suppose it's supposed to block off characters who aren't going commando. Not sure why the swimming blurb is there, but I'll just roll with it and add an extra blub here so context isn't all over the place--also makes use of the exposure checks.