Tweaked the double button controls in the options menu to hide the second button if they're only fed configuration info for one control.
Enabled the scroll bar for the options menu and tweaked the control layout slightly so that everything will in the slightly narrower horizontal space available with the scroll bar present.
Disabled mouse interaction with a bunch of the text elements on the options menu display to stop weird issues with scrolling the text when trying to use the scroll wheel to move the entire panel.
Added an option to hide the room text that overlays character portraits. Toggled in the options menu.
Added a control similar to the select-this-character-artist control that will temporarily toggle the display state of the room text- if normally visible, it will hide them until you take any game action or open any module (options, save/load/secondary output should all forcibly show it) do anything else. If normally hidden, it will show them until you do anything else.