Moved the Kelly bust that was in old gats' folder to adj's folder, and moved the [Embed]
Added an extra control up near the preference/room-hide control only visible in the image pack- if available, it'll show the "high res" version of the currently displayed bust. Which means we need to have a talk about how we're structuring the bust files/folders now before this turns into a clusterfuuuuuck.