Added Zaalt (alt. path to Uveto)
Added some placeholder stubs for if cybernetics/augs are present on a character
Added a quick means of checking if a certain friendly character was involved in combat
Added the alt. path to unlocking Uveto
Tweaked some existing Kara scenes to account for the minor changes the Icequeen path takes through them
Added a nude override to showKara
Added a stub to test if yammi is a crewmember
Unblocked a few of the rooms that I THINK the Ice Queen stuff needs on Uveto (the doc says Cafe, but Cafe isn't in the fucking main uveto doc- the closest thing is a bar, but the doc has no description for the bar...)
Changed the Uveto fly-to menu option to only be available whilst debug mode is enabled whilst we get more of the ground/content squared away
Added some markers to find the shitty placeholder text I rammed into room descripts so we don't accidently them.