Refactored the shopkeep system to used overridden methods on the creature objects rather than an ever-growing list of if-elses buried in items.as to determine what scene to execute when the player hits back in buy/sell menus
Added some goggles that give blind immunity, along with a new method on creature to check if a character is generally blind immune. Some of the blindness-application/avoidance texts have been modified to suit, but some have probably been missed and won't make sense or could stand to be properly extended to support the goggles.
Added an item flag that reduces accuracy of an attacker with the flag present on their accessory slot. (This should be expanded to support all slots...)
Added a climbing kit and hooked it up to some stuff on Myrellion to speed up travel & remove energy costs in places.
Added a bunch of misc items for Nerrasa
Added the group-welcome to Uveto that explains all of the dangers of the planet (moon) properly and expositions the player having a tracking device on-surface for mechanics reasons
Added a super slim version of Nerrasa to have a vendor that sells the PC heat belts for initial travel around Uveto; also sells a bunch of other misc stuff
Added room description for the interior of the geo survey station; theres a middle floor placeholdered but skipped until its actually used/needed by anything