Logic to avoid null clothing desc, bonus lactation line
Had a null variable error when the PC doesn't wear upper garments for pc.upperGarments. 
Added if statement to check for garments and if false it proceeds to my alternate output.
If PC is lactating without wearing upper garments it will display a bonus line. Not sure how well the bonus line works.
Else it continues as it did before.