Slammed in a means of force testing a scene exhibiting some issues
Reduced the width of the main text area slightly to give a little more padding against the scroll bar. Flash is scaling the game font non-uniformly if you resize the player window. Specifically, commas seem to consume a disproportiontely wider space, but flash doesn't attempt to reflow the text when the stage is scaled and not the /element/ containing the text (which is a correct optimization- reflowing text is heavy and by rights, it shouldn't need to be done). Scaling the stage up and not reflowing though, pushes any text after a comma slightly closer to the right, and in specific circumstances, its possible to "push" a character under the scrollbar.