Code cleanup and tweaks
Tweaked the Vagina class to return the tightness if it is between 0.5
and 1 since Mighty Tight can lower the minimum tightness to 0.5.

Tweaked the game.as to allow a vagina or ass to recover to the minimum
looseness. It previously stopped checking after the raw looseness
dropped below 2.

Changed the logic for exhibition decrease to be more concise and clear.

Removed a function from the creature class since it was only called in
one place and all the function did was return the value of another
function.