Wired up Hana's frozen recovery scene.
Fixed tryApplyUvetoColdDamage- if the player had 0 health prior to attempting to apply cold damage, no damage would be applied and the recovery process wouldn't start. Cold damage still has to be present, and it must still be applicable to the player even if they have 0 current health to trigger recovery.