Added an function package for list formatting - this will likely be renamed when I get around to ripping out the old list formatting stuff.
Minor fix for loading child data.
Modifications to the existing pregnancy handlers to add a duration API and a means of forcibly ending a pregnancy as though it had come to term. Allows the nursery systems to "finish" a pregnancy without the remaining stages being processed and triggering additional scenes, whilst also creating the children from the results of the pregnancy with the correct birth date.
Some of the pregnancy handlers and scenes may need further tweaking to ensure certain effects from the pregnancies are appropriately cleaned up, applied to the player, and that that flags that gate certain content are also correctly set.
Modified the item class to use a get/set property for descriptions, with a backing field. Items with dynamic descriptions can just override the getter to do whatever on a case-by-case basis, thus end-running around load-time data-order limitations.
Modified Siegwulfe to use the get description function to avoid a crash during load where the item can load in the players inventory before the backing character descriptions of Siegwulfe is loaded, hence breaking the ability to set a modified description string in the items constructor.
Further progress on the nursery content