Modified the nursery pregnancy wait code to ignore the egg trainers faux pregnancy.
Clean up the carry training eggs in the nursery better, invoking the additional effects and cleaning up the status. Wrapped some additional logic to hide output in the extra effects implenentation if called via the nursery ending a pregnancy.
Added an extra method to Creature to find pregnancies other than a specific type. Should probably be extended to support lists of potential ignores for future use.
Extended the main PregnancyManager to also support ignoring pregnancies of a specific type for finding the next slot due to complete.