More typos
The threesome stuff, as far as I can tell, still crashes due to the dick/knot stuff. Not sure if it will or won't crash in the current implementation with Fen's new code. Also the button for the threesome is buggy (shows up when I hadn't beaten Haley). Will work on dis, probably need someone to put out the fire for the former though.