Kazra, tweaks to the fertility priestess combat code
Added (yet) another hardlight-availability method that checks equipped, inventory AND the perk in a single go.
Added a provisional means to remove a creature from group combat. The UI code may not yet re-adapt to go from multiple->single enemy layouts yet however.
Added a means to get a creature from the group of hostiles with a specific class type. Handy for getting the "lead" character in a group fight.
Added items for the Essyra traders.
Blocked out most of the base glacial rift rooms + functions + fertility priestess encounter with the idol.