"Properly" fixed sticky formatting.
Sticky formatting is a weird edgecase where we have text output that inserts formatting tags, but not both the opening and closing tags at the same time. In most instances this is fine, but there are edge cases that causes this sticky formatting to break. In most scene body text, it seems to be okay to do this, but combat text output (even though it works identically as far as I can remember) seems to trigger this _all the time_. So, rather than stuffing text into the final output buffer each and every time we generate it, we now just mark the fact that the buffer has changed. At the start of a frame, we check if the buffers have changed and if they have, THEN we execute the functions that change the contents of the display object. In effect, this updates the display objects containing the main text body only once per player input, just before the content of the objects would be reflowed and rerendered for the changes. This should also marginally increase performance as any shit the display objects are doing in prep for a new frame will now only be done once; and we know it must have been doing something because of the sticky formatting.

Slimmed down the imports for the base Creature class by using wildcards and cleaned out a few "interfacey" functions in a couple of classes, instead prefering the Interface imports to provide implemenations.
Removed some exceptions from text output handling within the UI...