Minor Edits
* Fix for:
https://forum.fenoxo.com/threads/v0-7-164-shekka-eggs-pregnancy-time.13532/
* Add a num2Word() that changes integers to words like "a few", "a
couple" and has a couple parameters.
* Add a shortMinutes() that converts a minute value and lists out years,
months, weeks, days, hours, and/or minutes.
* Add getPregnancyTime() and getPregnancyTimeString() that will produce
the number or string, respectively, of a selected womb slot (or the womb
with the longest current duration if the slot is -1).
* Update a number of pregnancy blurbs that mention the time and had them
calculate the actual time (using the pregnancy's duration multiplier) to
output a more accurate message.