Updates to PlayerCharacter.as for NyreaEggCheck
Additional checks needed in PlayerCharacter.as for NyreaEgg fix.

StatusEffect would be readded if player had perk and ovipositor. This meant any other race with an ovipositor would trigger it. 

Additionally, eggs would continue to increase behind the scenes without this check.

The additional check for Nyrea_Cum limits this to Nyrea dicks only. This canonically makes sense because Nyrea eggs are inside the "cum".