The Pollen Dance Implementation
* Part 2 of the Plantation Expansion, by Nonesuch.
* If peaceful result from Plantation Quest, Steele can revisit Quinn and
talk to her about The Pollen Dance (a kind of fertility festival).
* Steele may choose to spectate the main event or actually participate.
* Quinn is able to become preggers with her own child (only once) with
slight variance based on whether or not the child is sired by Steele.
(Quinn may require pregnant bust art.)
* To ensure being the sire, Steele must agree to a 1-on-3 fight against
the other drones to prove worthiness. The fight is similar to the Lah
fight in that only zil weapons can be used.
* The fight has two new zil warrior types: the lean-yet-toned Spear Zil
and the trappy-ish Bow Zil. (Both may require bust art for normal and
nude variants.)
* Fight also has a passive pheromone lust gain at the end of the
player's turn. The strength of the attack is based on how many zil are
active on the field.
* Feel free to rebalance the fight if necessary. May need to disable
drones and special attacks to match the circumstances.
* If chosen to participate, Steele may also get gregnant during the
festivities if possessing a fertile vagina.
* Quinn's pregnancy should last for about 180 days. Steele may check up
on her periodically.
* If Steele is away for too long (beyond 245 days since the event),
Quinn may have the baby on her own (which means Steele will miss out on
some other in-between choices).
* Upon revisting Quinn, she will "name" her baby.
* If the baby is 365 days or older (toddler age), Steele may play with
him/her.
* The baby can only be a max age of 1825 days (5 years old) at the
moment.
* New status effect "Body Paint", acquired during the festival (paint
activity). Locks armor slot but allows tease advantage of 15%. Can be
washed away or wears off over time.
* "Penis Paint" can now also be washed away.
* New status "Full Belly", acquired during festival (feast activity).
Temporarily lowers physique and reflexes by 10% current value.
* New status "Pumped", acquired during festival (winning the jump rope
activity). Temporarily raises physique and aim by 15% of current.
* Updated stats to reflect changes.