Fixed formula in Creature.as
"plus a cylinder of full diameter and half height" means no need for that multiply by 2. And by the way, this - now correct - formula for double headed tip has exactly the same amount of volume as the simple, single headed, ordinary human penis tip. Check it out if you don't believe me! As an alternative, I'm going to create a new pull request in which I change this formula for a better one. Think about it!