New parsers (documentation in description)
-[char.rangedNoun]
-[char.gunNoun]
-[char.bowNoun]
These output the last word in a weapon's name because that's currently always the type of weapon it is and accounts for the odd "other" category of gun. To be used when the full name of a weapon would be overly much.

-ballsNounIsAre
Uses the ballsNoun parser and then adds either "is" or "are" based on whether the player has one ball or multiple.
-ballsNounSimpleIsAre
Same as above, but uses ballsNounSimple instead of ballsNoun as its basis