Silicone Expansion
* Minor forum bug/typo fixes.
* More Doctor Badger lab expansion content.
* Written by Natetheman223 and Kirorororo.
* The Doll Maker can now be fixed up (once) with a similar miniquest to
Nym-Foe.
* Play with individual Doll Maker arms or Fight it if non-Bimbo. Getting
dolled up leads to Adjatha's bad end.
* Strange Ball (brainwashing ball) can be used from the Badger's lab.
Basically resets intelligence and some proportions.
* Buy a silicone tank from Badger for the ship. Costs a lotta credits or
decide to carry silicone for a week.
* Combination of "Black Latex" and "Implant-tastic" perks will create
the "True Doll" perk.
* "True Doll" perk speeds up tightening of stretched orifices but can
completely ignore stretching if elasticity is high enough.
* "True Doll" increases sexiness by 5 and mentions the body being
"plastic" in the appearance screen.
* "True Doll" also adds a little elasticity to the character per day up
to an elasticity level of 5 max. The more times the character is
penetrated (cuntChange, buttChange, or cockHoleChange) during the day,
the higher the potential gain.
* May need to review any scenes that invovle urethra insertions
(sounding/docking) and add the cockHoleChange().