Small changes to keep some surprise
There are two changes in this commit, which are there to increase
the level of surprise relating to Amma and Zephyr.
The first change hides Amma on the first entry to New Texas since
she is "hiding" herself under Ogram's desk.
The second change hides the option to get milked by Zephyr in the
nursery until you've had kids with her and she actually visits it.